+++ research-hub-web/cypress/integration/basic-functionality.e2e.js
@@ -9,7 +9,7 @@ describe('ResearchHubs Basic Functionality', () => {
     });
 
     it('has title', () => {
-        cy.contains('Welcome to ResearchHub');
+        cy.contains('Welcome to the ResearchHub');
     });
 
     it('can browse by category', () => {
