+++ research-hub-web/src/app/app.component.html
@@ -49,7 +49,6 @@
                 <a routerLink="/home" class="hub-logo" ngClass.xs="hub-logo-xs">
                   <img [src]="'assets/hub-logo.svg'">
                 </a>
-
                 <span style="flex: 1 1 auto;"></span>
 
                 <button mat-icon-button fxHide.gt-sm routerLink="/search" matTooltip="Search">
@@ -98,7 +97,7 @@
                 </button>
 
                 <mat-menu #userMenu="matMenu" [overlapTrigger]="false">
-                  <div style="margin-right: 8px; margin-left: 8px;">
+                  <div class="user-detail-dropdown">
                     <img src="/assets/images/uoa-logo-blue.svg">
 
                     <mat-list class="user-menu-list">
@@ -113,9 +112,12 @@
                     </mat-list>
 
                     <div fxLayout="row" fxLayoutAlign="end">
-                      <button mat-raised-button style="color: black; background-color: white;"
-                        href="https://www.auckland.ac.nz/en/about-us/about-the-university/identity-and-access-management/single-sign-on/details.html">Sign 
-                        Out</button>
+                      <button 
+                        mat-raised-button 
+                        class="sign-out-button"
+                        href="https://www.auckland.ac.nz/en/about-us/about-the-university/identity-and-access-management/single-sign-on/details.html">
+                        Sign Out
+                      </button>
                     </div>
                   </div>
                 </mat-menu>
@@ -128,6 +130,7 @@
 
           <!-- <mat-drawer-container> -->
           <mat-sidenav-container fxFlex="noshrink">
+
             <!-- Content sidenav -->
             <mat-sidenav class="content-sidenav" mode="side" disableClose
               [opened]="appComponentService.isContentSidenavVisible" [fixedInViewport]="isContentSidenavFixed"
@@ -135,12 +138,15 @@
               <router-outlet name="contentSidenav" (activate)="setContentSidenavHasContent(true)"
                 (deactivate)="setContentSidenavHasContent(false)"></router-outlet>
             </mat-sidenav>
+
             <!-- Search bar  -->
             <mat-sidenav-content [@contentPushLeft]="appComponentService.isContentSidenavVisible">
-              <div class="site-padding" #content ngClass.xs="site-padding-xs" ngClass.sm="site-padding-sm"> 
+              <div #content> 
+
                  <!-- Banner -->
                  <div 
                   class="banner-container-home"
+                  [ngStyle]="{ 'background-image': coverImageUrl }"
                   *ngIf="showBanner">
                   <div class="banner-content-home">
                     <h1>{{ title }}</h1>
@@ -149,6 +155,7 @@
                 </div>  
                 <app-search-bar [categories]="categoryOptions" [category]="selectedCategory"
                   [searchText]="searchText"></app-search-bar>
+
                 <!-- Page content -->
                 <router-outlet></router-outlet>
               </div>
