+++ research-hub-web/src/app/components/articles/articles.component.html
@@ -1,8 +1,11 @@
 <!-- If we're displaying a collection of articles-->
-<app-collection-list 
+<div
     *ngIf="allArticles$ | async as allArticles" 
-    [collection]="allArticles">
-</app-collection-list>
+    class="site-padding margin-bottom top-padding" 
+    ngClass.xs="site-padding-xs margin-bottom-xs top-padding" 
+    ngClass.sm="site-padding-sm margin-bottom-sm top-padding">
+    <app-collection-list [collection]="allArticles"></app-collection-list>
+</div>
 
 <!-- If we're displaying an individual article -->
 <div 
@@ -13,7 +16,13 @@
     <ng-template #dynamicBanner>
         <div 
             class="banner-container"
-            [style]="article.displayBanner ? article.banner ? 'background: url(' + article.banner.url + ') no-repeat' : 'background: url(../../../assets/images/1483_UoA_13Oct09_1680x220_BW.jpg)' : article.viewType == 'View 2' ? 'display: none;' : ''">
+            [style]="article.displayBanner ? 
+                        article.banner ? 
+                            'background: url(' + article.banner.url + ') no-repeat' :
+                            'background: url(../../../assets/images/1483_UoA_13Oct09_1680x220_BW.jpg)' :
+                        article.viewType == 'View 2' ? 
+                            'display: none' : 
+                            ''">
         </div>  
     </ng-template>
 
@@ -21,7 +30,10 @@
         <div 
             id="breadcrumbs" 
             *ngIf="!!parentSubHubs" 
-            ngClass.xs="breadcrumbs-mobile">
+            class="site-padding top-padding"
+            ngClass.sm="site-padding-sm top-padding"
+            ngClass.xs="site-padding-xs top-padding"
+            ngClass.xs="breadcrumbs-mobile site-padding-xs">
             <span 
                 class="breadcrumb-subhub" 
                 *ngFor="let subHub of parentSubHubs.slice().reverse()">
@@ -40,7 +52,7 @@
             class="article-title" 
             ngClass.sm="article-title-sm" 
             ngClass.xs="article-title-xs"
-            [style.color]="article.displayBanner ? 'white' : ''">
+            [style.color]="article.displayBanner ? article.viewType == 'View 1' ? 'white' : '' : ''">
             {{ article.title }}
         </h1>
 
@@ -48,14 +60,13 @@
             class="article-summary" 
             ngClass.sm="article-summary-sm" 
             ngClass.xs="article-summary-xs"
-            [style.color]="article.displayBanner ? 'white' : ''">
+            [style.color]="article.displayBanner ? article.viewType == 'View 1' ? 'white' : '' : ''">
             {{ article.summary }}
         </h3>
     </ng-template>
 
     <ng-template #articleBody>
         <mat-card-content 
-            style="margin-bottom: 50px;" 
             [innerHTML]="article.body.json | richTextToHTML">
         </mat-card-content>
     </ng-template>
@@ -181,8 +192,6 @@
     </ng-template>
     
     
-    
-    
     <!-- View Switcher -->
     <div [ngSwitch]="article.viewType"> 
 
@@ -192,7 +201,10 @@
             <!-- Dynamic Banner -->
             <div 
                 class="banner-container"
-                [style]="article.displayBanner ? article.banner ? 'background: url(' + article.banner.url + ') no-repeat' : 'background: url(../../../assets/images/1483_UoA_13Oct09_1680x220_BW.jpg)' : article.viewType == 'View 2' ? 'display: none;' : ''">
+                [style]="article.displayBanner ? 
+                            article.banner ? 
+                                'background: url(' + article.banner.url + ') no-repeat' :
+                                'background: url(../../../assets/images/1483_UoA_13Oct09_1680x220_BW.jpg)' : ''">
                 <div 
                     class="banner-content" 
                     *ngIf="article.viewType == 'View 1'">
@@ -204,6 +216,9 @@
             <ng-container *ngTemplateOutlet="breadcrumbs"></ng-container>
 
             <div 
+                class="site-padding top-padding" 
+                ngClass.xs="site-padding-xs top-padding" 
+                ngClass.sm="site-padding-sm top-padding"
                 fxLayout="row wrap" 
                 fxLayout.sm="column wrap"
                 fxLayout.xs="column wrap"   
@@ -259,47 +274,46 @@
             <ng-container *ngTemplateOutlet="breadcrumbs"></ng-container>
 
             <div 
+                class="site-padding"
                 fxLayout="row wrap" 
                 fxLayout.sm="column wrap"
                 fxLayout.xs="column wrap"   
+                fxLayoutAlign="space-between"
                 fxLayoutGap="5%">
 
                 <!-- Article Title and Summary -->
-                <div fxFlex="70%">
-                    <ng-container *ngTemplateOutlet="titleSummary"></ng-container>
+                    <div fxFlex="70%">
+                        <ng-container *ngTemplateOutlet="titleSummary"></ng-container>
 
-                    <hr style="margin: 25px 0px;"/>
+                        <hr style="margin: 25px 0px;"/>
 
-                    <!-- Article Body -->
-                    <ng-container *ngTemplateOutlet="articleBody"></ng-container>
+                        <!-- Article Body -->
+                        <ng-container *ngTemplateOutlet="articleBody"></ng-container>
 
-                </div>
+                    </div>
 
-                <!-- Extra Links -->
-                <div 
-                    fxLayout="column" 
-                    fxFlex="25%">
+                    <!-- Extra Links -->
+                    <div 
+                        fxLayout="column" 
+                        fxFlex="25%">
 
-                    <!-- Official Documents -->
-                    <ng-container *ngTemplateOutlet="officialDocuments"></ng-container>
+                        <!-- Official Documents -->
+                        <ng-container *ngTemplateOutlet="officialDocuments"></ng-container><div><br/></div>
 
-                    <!-- Contacts -->
-                    <ng-container *ngTemplateOutlet="relatedContacts"></ng-container>
+                        <!-- Contacts -->
+                        <ng-container *ngTemplateOutlet="relatedContacts"></ng-container><div><br/></div>
 
-                    <!-- Organizations -->
-                    <ng-container *ngTemplateOutlet="relatedOrganisations"></ng-container>
+                        <!-- Organizations -->
+                        <ng-container *ngTemplateOutlet="relatedOrganisations"></ng-container><div><br/></div>
 
-                </div>
+                        <!-- Related Items -->
+                        <ng-container *ngTemplateOutlet="relatedItems"></ng-container><div><br/></div>
+
+                        <!-- Keywords -->
+                        <ng-container *ngTemplateOutlet="keywords"></ng-container>
+
+                    </div>
             </div>
-
-            <!-- Keywords -->
-            <ng-container *ngTemplateOutlet="keywords"></ng-container>
-
-            <br/>
-
-            <!-- Related Items -->
-            <ng-container *ngTemplateOutlet="relatedItems"></ng-container>
-
         </div>
     </div>
