+++ research-hub-web/src/app/components/articles/articles.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { pluck, map, flatMap } from 'rxjs/operators';
+import { pluck, map, flatMap, tap } from 'rxjs/operators';
 import { ActivatedRoute } from '@angular/router';
-import { AppComponentService } from '../../app.component.service';
+import { AppComponentService } from '@app/app.component.service';
 import {
   AllArticlesGQL,
   GetArticleBySlugGQL,
@@ -40,14 +40,14 @@ export class ArticlesComponent implements OnInit {
      */
     this.slug = this.route.snapshot.params.slug || this.route.snapshot.data.slug;
 
-
     /**
      * If this.slug is defined, we're loading an individual article,
      * therefore run the corresponding query. If not, return all articles.
      */
     if (!!this.slug) {
       this.getArticleBySlug(this.slug).subscribe(data => {
-        this.article$ = this.getArticleByID(data.sys.id);
+        this.article$ = this.getArticleByID(data.sys.id)
+          .pipe(tap(res => this.appComponentService.setTitle(res.title)))
       });
       this.parentSubHubs = await this.cerGraphQLService.getParentSubHubs(this.slug);
     } else {
