+++ research-hub-web/src/app/components/home/browse/browse.component.html
@@ -1,36 +1,69 @@
-<h2 *ngIf="embedded">Browse categories</h2>
-<mat-grid-list [cols]="numCols" gutterSize="1em">
-  <mat-grid-tile class="browse-tile" *ngFor="let item of categoryOptions;" [colspan]="1" [rowspan]="1"
-    [routerLink]="['/search']" [queryParams]="{categoryId: item.id}" [matTooltip]="getCategoryDescription(item.id)"
-    [matTooltipShowDelay]="500">
-    <div fxLayout="column" fxLayoutAlign="center center" class="tile-content">
-      <span class="tile-text">{{item.name}}</span>
-      <mat-icon 
-        class="tile-icon"
-        ngClass.lg="tile-icon-lg"
-        ngClass.md="tile-icon-md"
-        ngClass.sm="tile-icon-sm"
-        ngClass.xs="tile-icon-xs">
-          {{item.icon}}
-        </mat-icon>
-    </div>
-  </mat-grid-tile>
-</mat-grid-list>
-
-<h2 *ngIf="embedded" style="margin-top: 2vh;">NEW - Browse GraphQL Categories</h2>
-<mat-grid-list [cols]="numCols" gutterSize="1em">
-  <mat-grid-tile class="browse-tile-new" *ngFor="let item of categoryOptionsGQL;" [colspan]="1"
-    [rowspan]="1" [routerLink]="item.url" [matTooltip]="getCategoryDescription(item.id)" [matTooltipShowDelay]="500">
-    <div fxLayout="column" fxLayoutAlign="center center" class="tile-content">
-      <span class="tile-text">{{item.name}}</span>
-      <mat-icon 
-        class="tile-icon"
-        ngClass.lg="tile-icon-lg"
-        ngClass.md="tile-icon-md"
-        ngClass.sm="tile-icon-sm"
-        ngClass.xs="tile-icon-xs">
-          {{item.icon}}
-      </mat-icon>
-    </div>
-  </mat-grid-tile>
-</mat-grid-list>
+<div>
+  <h2 *ngIf="embedded">
+    Browse categories
+  </h2>
+  <mat-grid-list 
+    [cols]="numCols" 
+    gutterSize="1em">
+    <mat-grid-tile 
+      class="browse-tile" 
+      *ngFor="let item of categoryOptions;" 
+      [colspan]="1" 
+      [rowspan]="1"
+      [routerLink]="['/search']" 
+      [queryParams]="{categoryId: item.id}" 
+      [matTooltip]="getCategoryDescription(item.id)"
+      [matTooltipShowDelay]="500">
+      <div fxLayout="column" 
+        fxLayoutAlign="center center" 
+        class="tile-content">
+        <span class="tile-text">
+          {{ item.name }}
+        </span>
+        <mat-icon 
+          class="tile-icon"
+          ngClass.lg="tile-icon-lg"
+          ngClass.md="tile-icon-md"
+          ngClass.sm="tile-icon-sm"
+          ngClass.xs="tile-icon-xs">
+            {{ item.icon }}
+          </mat-icon>
+      </div>
+    </mat-grid-tile>
+  </mat-grid-list>
+  <div
+    class="top-padding"
+    ngClass.xs="top-padding" 
+    ngClass.sm="top-padding">
+    <h2 *ngIf="embedded">
+        NEW - Browse GraphQL Categories
+    </h2>
+    <mat-grid-list 
+      [cols]="numCols" 
+      gutterSize="1em">
+      <mat-grid-tile 
+        class="browse-tile-new" 
+        *ngFor="let item of categoryOptionsGQL;" 
+        [colspan]="1"
+        [rowspan]="1" 
+        [routerLink]="item.url" 
+        [matTooltip]="getCategoryDescription(item.id)" 
+        [matTooltipShowDelay]="500">
+        <div 
+          fxLayout="column" 
+          fxLayoutAlign="center center" 
+          class="tile-content">
+          <span class="tile-text">{{item.name}}</span>
+          <mat-icon 
+            class="tile-icon"
+            ngClass.lg="tile-icon-lg"
+            ngClass.md="tile-icon-md"
+            ngClass.sm="tile-icon-sm"
+            ngClass.xs="tile-icon-xs">
+              {{item.icon}}
+          </mat-icon>
+        </div>
+      </mat-grid-tile>
+    </mat-grid-list>
+  </div>
+</div>
