+++ research-hub-web/src/app/components/home/home.component.html
@@ -1,11 +1,31 @@
-
-<div fxLayout="column" fxLayoutGap="2.8em" style="margin-top: 2.8em;">
+<div 
+  class="margin-bottom"
+  ngClass.xs="margin-bottom-xs" 
+  ngClass.sm="margin-bottom-sm"
+  fxLayout="column" 
+  fxLayoutGap="2.8em" 
+  style="margin-top: 2.8em;">
   <app-research-activity></app-research-activity>
-  <div fxFlex="noshrink" fxLayout="row" fxLayout.xs="column" fxLayoutGap="7%">
-    <div fxFlex="65" fxFlex.xs="noshrink">
-      <app-browse [embedded]="true" [maxCols]="3"></app-browse>
+  <div 
+    class="site-padding-2" 
+    ngClass.xs="site-padding-xs" 
+    ngClass.sm="site-padding-sm"
+    fxFlex="noshrink" 
+    fxLayout="row" 
+    fxLayout.xs="column" 
+    fxLayoutGap="7%">
+    <div 
+      fxFlex="65" 
+      fxFlex.xs="noshrink">
+      <app-browse 
+        [embedded]="true" 
+        [maxCols]="3">
+      </app-browse>
     </div>
-    <div fxFlex="35" fxFlex.xs="noshrink" fxLayoutGap="5%">
+    <div 
+      fxFlex="35" 
+      fxFlex.xs="noshrink" 
+      fxLayoutGap="5%">
       <app-featured></app-featured>
     </div>
   </div>
