+++ research-hub-web/src/app/components/home/home.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { SearchBarService } from '../search-bar/search-bar.service';
 import { FeaturedComponent } from './featured/featured.component';
 import { BrowseComponent } from './browse/browse.component';
+import { CategoryId } from '@app/global/global-variables';
 
 
 @Component({
@@ -15,19 +16,6 @@ export class HomeComponent implements OnInit {
   }
 
   ngOnInit() {
-    enum CategoryId {
-      All = 1,
-      Support,
-      Equipment,
-      Training,
-      Software,
-      Facilities,
-      Guide,
-      Person,
-      Policies,
-      Articles,
-      SubHubs
-    }
     this.searchBarService.setSearchText('');
     this.searchBarService.setCategory(CategoryId.All);
   }
