+++ research-hub-web/src/app/components/home/research-activity/research-activity.component.html
@@ -1,16 +1,21 @@
-<h2>Explore research activities</h2>
+<div
+  class="site-padding-2" 
+  ngClass.xs="site-padding-xs" 
+  ngClass.sm="site-padding-sm">
+  <h2>Explore research activities</h2>
 
-<mat-chip-list>
-  <div fxLayout="row wrap" fxLayout.lt-sm="column" fxLayoutAlign="space-between" fxLayoutAlign.lt-lg="flex-start"
-    style="width: 100%">
-    <div *ngFor="let activity of researchActivityOptions" [ngClass]="activity.className">
-      <!-- Need to specify class for containing div also in order to use custom themes on material components -->
-      <mat-chip selected [routerLink]="['/search']" [queryParams]="getQueryParams(activity)"
-        [ngClass]="[activity.className, 'ra-chip']" [ngClass.lt-md]="[activity.className, 'ra-chip', 'tablet-view']"
-        [ngClass.lt-sm]="[activity.className, 'ra-chip', 'mobile-view']"
-        (click)="analyticsService.trackUserExperience('Home page research activities', 'click')">
-        {{activity.name}}
-      </mat-chip>
+  <mat-chip-list>
+    <div fxLayout="row wrap" fxLayout.lt-sm="column" fxLayoutAlign="space-between" fxLayoutAlign.lt-lg="flex-start"
+      style="width: 100%">
+      <div *ngFor="let activity of researchActivityOptions" [ngClass]="activity.className">
+        <!-- Need to specify class for containing div also in order to use custom themes on material components -->
+        <mat-chip selected [routerLink]="['/search']" [queryParams]="getQueryParams(activity)"
+          [ngClass]="[activity.className, 'ra-chip']" [ngClass.lt-md]="[activity.className, 'ra-chip', 'tablet-view']"
+          [ngClass.lt-sm]="[activity.className, 'ra-chip', 'mobile-view']"
+          (click)="analyticsService.trackUserExperience('Home page research activities', 'click')">
+          {{activity.name}}
+        </mat-chip>
+      </div>
     </div>
-  </div>
-</mat-chip-list>
+  </mat-chip-list>
+</div>
