+++ research-hub-web/src/app/components/org-unit-details/org-unit-details.component.ts
@@ -6,12 +6,7 @@ import { Location } from '@angular/common';
 import { AnalyticsService } from '@services/analytics.service';
 import { ListItem } from '@model/ListItem';
 import { AppComponentService } from '@app/app.component.service';
-
-enum RoleTypeId {
-  ServiceOwner = 1,
-  CopyOwner,
-  UserSupport
-}
+import { RoleTypeId } from '@app/global/global-variables';
 
 @Component({
   selector: 'app-org-unit-details',
