+++ research-hub-web/src/environments/environment.ts
@@ -29,4 +29,5 @@ export const environment = {
       return []
     },
   },
+  
 }
