+++ research-hub-web/src/themes.scss
@@ -4,6 +4,17 @@
 // Include the common styles for Angular Material. Be sure that you only ever include this mixin once!
 @include mat-core();
 
+// University of Auckland colors from Web Style Guide 
+$dark-blue: #00467F;
+$light-blue: #009AC7;
+$silver: #8D9091;
+$dark-grey: #333333;
+$mid-grey: #404040;
+$body-grey: #4A4A4C;
+$silver-lining: #BEC3C4;
+$input-grey: #D9D9D9;
+$light-grey: #F2F2F2;
+
 // Material design colors generated with: http://mcg.mbitson.com/
 $hub-teal: (
   50 : #e1f2f4,
@@ -145,12 +156,12 @@ $hub-app-theme: mat-light-theme($hub-app-primary, $hub-app-accent, $hub-app-warn
 @include angular-material-theme($hub-app-theme);
 
 // Site wide measurements
-$width-page-max: 85vw;
-$offset-xs: 1em;
+$width-page-max: 100vw;
 $offset-gt-xs: 2em;
 $height-footer: 12em;
 $height-footer-xs: 16em;
-$padding-site: 3em;
+$padding-site: 7vw;
+$padding-site-top: 3vh;
 
 // Site wide colors
 $color-primary: map-get($hub-app-primary, 500);
